Refactor: #27 예약 관련 패키지명 booking으로 변경
기존 패키지 ticket에서 booking으로 변경했다.
