+++ src/main/java/com/sparta/parknav/booking/repository/CarRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.parknav.ticket.repository;
+package com.sparta.parknav.booking.repository;
 
-import com.sparta.parknav.ticket.entity.Car;
+import com.sparta.parknav.booking.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CarRepository extends JpaRepository<Car, Long> {
