+++ src/main/java/com/sparta/parknav/management/service/MgtService.java
@@ -13,8 +13,8 @@ import com.sparta.parknav.management.dto.request.CarNumRequestDto;
 import com.sparta.parknav.management.dto.response.CarInResponseDto;
 import com.sparta.parknav.management.dto.response.CarOutResponseDto;
 import com.sparta.parknav.management.dto.response.ParkMgtResponseDto;
-import com.sparta.parknav.ticket.entity.ParkBookingInfo;
-import com.sparta.parknav.ticket.repository.ParkBookingInfoRepository;
+import com.sparta.parknav.booking.entity.ParkBookingInfo;
+import com.sparta.parknav.booking.repository.ParkBookingInfoRepository;
 import com.sparta.parknav.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
